pyproject.toml
[project]
name = "engram"
version = "0.1.0dev"
description = "Self-improving procedural memory for Claude Code"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-frontmatter>=1.1",
    "filelock>=3.12",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
llm = [
    "anthropic>=0.30",
]

[project.scripts]
engram = "engram.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/engram"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["engram"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = "frontmatter.*"
ignore_missing_imports = true
